requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.5
numpy>=1.21.0
flask>=2.0.0
face-recognition>=1.3.0
dlib>=19.22.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
pillow>=8.3.1
python-dotenv>=0.19.0
SQLAlchemy>=1.4.23
flask-sqlalchemy>=2.5.1
flask-login>=0.5.0
flask-wtf>=0.15.1 